MAINT decompose implementation
